+++ src/main/clojure/com/github/clojure_repl/intellij/action/eval.clj
@@ -12,25 +12,27 @@
   (:import
    [com.intellij.openapi.actionSystem CommonDataKeys]
    [com.intellij.openapi.actionSystem AnActionEvent]
-   [com.intellij.openapi.editor Editor]))
+   [com.intellij.openapi.editor Editor]
+   [com.intellij.openapi.vfs VirtualFile]))
 
 (set! *warn-on-reflection* true)
 
 (defn load-file-action [^AnActionEvent event]
-  (when-let [editor ^Editor (.getData event CommonDataKeys/EDITOR_EVEN_IF_INACTIVE)]
-    (let [project (.getProject editor)]
-      (if (db/get-in project [:current-nrepl :session-id])
-        (tasks/run-background-task!
-         (.getProject editor)
-         "REPL: Loading file"
-         (fn [_indicator]
-           (let [{:keys [status err]} (nrepl/load-file project editor)]
-             (app-manager/invoke-later!
-              {:invoke-fn (fn []
-                            (if (and (contains? status "eval-error") err)
-                              (ui.hint/show-repl-error :message err :editor editor)
-                              (ui.hint/show-repl-info :message (str "Loaded file " (.getPath (.getVirtualFile editor))) :editor editor)))}))))
-        (ui.hint/show-error :message "No REPL connected" :editor editor)))))
+  (when-let [virtual-file ^VirtualFile (.getData event CommonDataKeys/VIRTUAL_FILE)]
+    (when-let [editor ^Editor (.getData event CommonDataKeys/EDITOR_EVEN_IF_INACTIVE)]
+      (let [project (.getProject editor)]
+        (if (db/get-in project [:current-nrepl :session-id])
+          (tasks/run-background-task!
+           (.getProject editor)
+           "REPL: Loading file"
+           (fn [_indicator]
+             (let [{:keys [status err]} (nrepl/load-file project editor virtual-file)]
+               (app-manager/invoke-later!
+                {:invoke-fn (fn []
+                              (if (and (contains? status "eval-error") err)
+                                (ui.hint/show-repl-error :message err :editor editor)
+                                (ui.hint/show-repl-info :message (str "Loaded file " (.getPath ^VirtualFile virtual-file)) :editor editor)))}))))
+          (ui.hint/show-error :message "No REPL connected" :editor editor))))))
 
 (defn eval-last-sexpr-action [^AnActionEvent event]
   ;; TODO change for listeners here or a better way to know which repl is related to current opened file