+++ src/main/clojure/com/github/clojure_repl/intellij/configuration/settings_editor/remote.clj
@@ -1,0 +1,79 @@
+(ns com.github.clojure-repl.intellij.configuration.settings-editor.remote
+  (:gen-class
+   :state state
+   :init init
+   :name com.github.clojure_repl.intellij.configuration.settings_editor.Remote
+   :extends com.intellij.openapi.options.SettingsEditor)
+  (:require
+   [seesaw.core :as seesaw]
+   [seesaw.mig :as mig])
+  (:import
+   [com.intellij.openapi.project Project ProjectManager]
+   [com.intellij.ui IdeBorderFactory]
+   [javax.swing JRadioButton JTextField]))
+
+(set! *warn-on-reflection* false)
+
+(defn ^:private mode-id-key [repl-mode]
+  (->> (seesaw/selection repl-mode)
+       (seesaw/id-of)))
+
+(defn ^:private build-editor-ui []
+  (let [repl-mode-group (seesaw/button-group)
+        panel (mig/mig-panel
+               :border (IdeBorderFactory/createTitledBorder "NREPL connection")
+               :items [[(seesaw/radio :text "Manual"
+                                      :id :manual
+                                      :group repl-mode-group
+                                      :mnemonic \M
+                                      :selected? true) "wrap"]
+                       [(seesaw/label "Host") ""]
+                       [(seesaw/text :id :nrepl-host
+                                     :columns 20) "wrap"]
+                       [(seesaw/label "Port") ""]
+                       [(seesaw/text :id :nrepl-port
+                                     :columns 8) "wrap"]
+                       [(seesaw/radio :text "Read from repl file"
+                                      :id :repl-file
+                                      :group repl-mode-group
+                                      :mnemonic \R) "wrap"]
+                       [(seesaw/label "Project") ""]
+                       [(seesaw/combobox :id :project
+                                         :model (->> (ProjectManager/getInstance)
+                                                     .getOpenProjects
+                                                     (map #(.getName ^Project %)))) "wrap"]])]
+    (seesaw/listen repl-mode-group :action
+                   (fn [_e]
+                     (let [mode-key (mode-id-key repl-mode-group)
+                           manual? (= mode-key :manual)]
+                       (.setEnabled ^JTextField (seesaw/select panel [:#nrepl-host]) manual?)
+                       (.setEnabled ^JTextField (seesaw/select panel [:#nrepl-port]) manual?))))
+
+    panel))
+
+(defn -init []
+  [[] (atom (build-editor-ui))])
+
+(defn -createEditor [this]
+  @(.state this))
+
+(defn ^:private manual? [editor]
+  (.isSelected ^JRadioButton (seesaw/select editor [:#manual])))
+
+(defn -applyEditorTo [this configuration]
+  (let [ui @(.state this)
+        host (seesaw/text (seesaw/select ui [:#nrepl-host]))
+        project-path (seesaw/text (seesaw/select ui [:#project]))
+        mode (if (manual? ui) :manual-config :file-config)]
+    (.setNreplHost (.getOptions configuration) host)
+    (.setMode (.getOptions configuration) (name mode))
+    (when-let [nrepl-port (parse-long (seesaw/text (seesaw/select ui [:#nrepl-port])))]
+      (.setNreplPort (.getOptions configuration) (str nrepl-port))
+      (.setProject (.getOptions configuration) project-path))))
+
+(defn -resetEditorFrom [this configuration]
+  (let [ui @(.state this)
+        options (.getOptions configuration)]
+    (seesaw/text! (seesaw/select ui [:#nrepl-host]) (.getNreplHost options))
+    (seesaw/text! (seesaw/select ui [:#nrepl-port]) (.getNreplPort options))
+    (seesaw/text! (seesaw/select ui [:#project]) (.getProject options))))